created abstract class Component.js
